requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via instructor
aiosignal==1.4.0
    # via aiohttp
aistudio-sdk==0.3.8
    # via paddlex
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
asgiref==3.10.0
    # via
    #   daphne
    #   django
attrs==25.4.0
    # via
    #   aiohttp
    #   service-identity
    #   twisted
autobahn==25.10.2
    # via daphne
automat==25.4.16
    # via twisted
bce-python-sdk==0.9.50
    # via aistudio-sdk
billiard==4.2.2
    # via celery
celery==5.5.3
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
cfgv==3.4.0
    # via pre-commit
chardet==5.2.0
    # via paddlex
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   aistudio-sdk
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   typer
    #   typer-slim
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorlog==6.10.1
    # via paddlex
constantly==23.10.4
    # via twisted
cron-descriptor==2.0.6
    # via django-celery-beat
cryptography==46.0.3
    # via
    #   autobahn
    #   jwcrypto
    #   pyopenssl
    #   service-identity
daphne==4.2.1
    # via -r requirements.in
diskcache==5.6.3
    # via instructor
distlib==0.4.0
    # via virtualenv
distro==1.9.0
    # via openai
django==5.2.7
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-oauth-toolkit
    #   django-timezone-field
    #   djangorestframework
    #   graphene-django
django-celery-beat==2.8.1
    # via -r requirements.in
django-oauth-toolkit==3.1.0
    # via -r requirements.in
django-timezone-field==7.1
    # via
    #   -r requirements.in
    #   django-celery-beat
djangorestframework==3.16.1
    # via -r requirements.in
docstring-parser==0.17.0
    # via instructor
filelock==3.20.0
    # via
    #   huggingface-hub
    #   modelscope
    #   paddlex
    #   virtualenv
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via huggingface-hub
future==1.0.0
    # via bce-python-sdk
graphene==3.4.3
    # via graphene-django
graphene-django==3.2.3
    # via -r requirements.in
graphql-core==3.2.7
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==3.2.0
    # via
    #   graphene
    #   graphene-django
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   huggingface-hub
    #   openai
    #   paddlepaddle
huggingface-hub==1.1.2
    # via paddlex
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
identify==2.6.15
    # via pre-commit
idna==3.11
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
    #   twisted
    #   yarl
imagesize==1.4.1
    # via paddlex
incremental==24.7.2
    # via twisted
instructor==1.13.0
    # via -r requirements.in
jinja2==3.1.6
    # via instructor
jiter==0.11.1
    # via
    #   instructor
    #   openai
jwcrypto==1.5.6
    # via django-oauth-toolkit
kombu==5.5.4
    # via celery
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
modelscope==1.31.0
    # via paddlex
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
networkx==3.5
    # via paddlepaddle
nodeenv==1.9.1
    # via pre-commit
numpy==2.3.4
    # via
    #   -r requirements.in
    #   opencv-contrib-python
    #   opencv-python
    #   opt-einsum
    #   paddlepaddle
    #   paddlex
    #   pandas
    #   shapely
oauthlib==3.3.1
    # via django-oauth-toolkit
openai==2.7.1
    # via
    #   -r requirements.in
    #   instructor
opencv-contrib-python==4.10.0.84
    # via paddlex
opencv-python==4.11.0.86
    # via -r requirements.in
opt-einsum==3.3.0
    # via paddlepaddle
packaging==25.0
    # via
    #   huggingface-hub
    #   kombu
    #   paddlex
paddleocr==3.3.1
    # via -r requirements.in
paddlepaddle==3.2.1
    # via -r requirements.in
paddlex[ocr-core]==3.3.8
    # via paddleocr
pandas==2.3.3
    # via paddlex
pillow==12.0.0
    # via
    #   paddlepaddle
    #   paddlex
platformdirs==4.5.0
    # via virtualenv
pre-commit==4.3.0
    # via instructor
prettytable==3.16.0
    # via
    #   aistudio-sdk
    #   paddlex
promise==2.3
    # via graphene-django
prompt-toolkit==3.0.52
    # via click-repl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.0
    # via paddlepaddle
psutil==7.1.3
    # via aistudio-sdk
psycopg2-binary==2.9.11
    # via -r requirements.in
py-cpuinfo==9.0.0
    # via paddlex
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pyclipper==1.3.0.post6
    # via paddlex
pycparser==2.23
    # via cffi
pycryptodome==3.23.0
    # via bce-python-sdk
pydantic==2.12.4
    # via
    #   instructor
    #   openai
    #   paddlex
pydantic-core==2.41.5
    # via
    #   instructor
    #   pydantic
pygments==2.19.2
    # via rich
pyopenssl==25.3.0
    # via twisted
pypdfium2==5.0.0
    # via paddlex
python-bidi==0.6.7
    # via paddlex
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   graphene
    #   pandas
python-dotenv==1.2.1
    # via -r requirements.in
python3-logstash==0.4.80
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   paddleocr
    #   paddlex
    #   pre-commit
redis==7.0.1
    # via -r requirements.in
requests==2.32.5
    # via
    #   aistudio-sdk
    #   django-oauth-toolkit
    #   instructor
    #   modelscope
    #   paddleocr
    #   paddlex
rich==14.2.0
    # via
    #   instructor
    #   typer
ruamel-yaml==0.18.16
    # via paddlex
ruamel-yaml-clib==0.2.14
    # via ruamel-yaml
safetensors==0.6.2
    # via paddlepaddle
service-identity==24.2.0
    # via twisted
shapely==2.1.2
    # via paddlex
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
six==1.17.0
    # via
    #   bce-python-sdk
    #   promise
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlparse==0.5.3
    # via django
tenacity==9.1.2
    # via instructor
text-unidecode==1.3
    # via graphene-django
tqdm==4.67.1
    # via
    #   aistudio-sdk
    #   huggingface-hub
    #   modelscope
    #   openai
twisted[tls]==25.5.0
    # via daphne
txaio==25.9.2
    # via autobahn
ty==0.0.1a25
    # via instructor
typer==0.20.0
    # via instructor
typer-slim==0.20.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   cron-descriptor
    #   graphene
    #   huggingface-hub
    #   jwcrypto
    #   openai
    #   paddleocr
    #   paddlepaddle
    #   paddlex
    #   pydantic
    #   pydantic-core
    #   twisted
    #   typer
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via
    #   django-celery-beat
    #   kombu
    #   pandas
ujson==5.11.0
    # via paddlex
urllib3==2.5.0
    # via
    #   modelscope
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.35.4
    # via pre-commit
wcwidth==0.2.14
    # via
    #   prettytable
    #   prompt-toolkit
yarl==1.22.0
    # via aiohttp
zope-interface==8.1
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
